requirements.txt
aiogram==3.10.1
python-dotenv